rag/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai==1.35.1
langchain==0.2.5
langchain-openai==0.1.9
langchain-community==0.2.5
faiss-cpu>=1.9.0
numpy==1.26.4
pydantic==2.7.4